:sparkles: Implement TemplateLookup interface to replace opaque lambda
We previously used lambdas as functions to call to lookup a template by name (then from an FS or from a directory straight). This was a bit opaque, and ultimately a bit unflexible.

Instead of definining the lambda signature as the TemplateLookup type,
this introduces a proper interface {}, which allows to pave the
road for a reload functionality ,which would allow us to reload the content files
 for sqleton dynamically from a refresh endpoint.